Refactor backend code and add javadoc to backend code
