+++ src/main/java/org/codedefenders/assistant/GPTObjects/GPTCompletion.java
@@ -1,7 +1,14 @@
-package org.codedefenders.smartassistant.GPTObjects;
+package org.codedefenders.assistant.GPTObjects;
 
 import java.util.List;
 
+/**
+ * This class represents a chat completion object for the GPT API. Specifications about chat completion properties are
+ * available at
+ * <a href="https://platform.openai.com/docs/api-reference/chat/object">
+ *     https://platform.openai.com/docs/api-reference/chat/object
+ * </a>
+ */
 public class GPTCompletion {
         private String id;
         private String object;
