+++ src/main/java/org/codedefenders/assistant/entities/AssistantType.java
@@ -1,0 +1,8 @@
+package org.codedefenders.assistant.entities;
+
+/**
+ * This enum represents the type of smart assistant assigned to a user.
+ */
+public enum AssistantType {
+    NONE, NOT_GUIDED
+}
