+++ src/main/java/org/codedefenders/assistant/entities/AssistantUserSettingsEntity.java
@@ -1,18 +1,28 @@
-package org.codedefenders.model;
+package org.codedefenders.assistant.entities;
 
+/**
+ * This class represents a user settings entity in the database. A user settings entity stores user's basic information,
+ * the total number of questions asked by a user, the remaining amount of questions that a user can ask and the type of
+ * assistant assigned to the user.
+ */
 public class AssistantUserSettingsEntity {
 
     private transient Integer id;
     private Integer userId;
-    private String username;
-    private String email;
-    private Integer questionsNumber;
-    private Integer remainingQuestions;
-    private SmartAssistantType assistantType;
+    private final String username;
+    private final String email;
+    private final Integer questionsNumber;
+    private final Integer remainingQuestions;
+    private final AssistantType assistantType;
+
+    /**
+     * This field stores the amount of questions to be added (if positive) or removed (if negative) from the remaining
+     * questions of the user. It is not stored in the database.
+     */
     private Integer remainingQuestionsDelta;
 
     public AssistantUserSettingsEntity(Integer id, Integer userId, String username, String email, Integer questionsNumber,
-                                       Integer remainingQuestions, SmartAssistantType assistantType) {
+                                       Integer remainingQuestions, AssistantType assistantType) {
         this.id = id;
         this.userId = userId;
         this.username = username;
@@ -39,20 +49,24 @@ public class AssistantUserSettingsEntity {
         this.userId = userId;
     }
 
-    public SmartAssistantType getAssistantType() {
-        return assistantType;
+    public String getUsername() {
+        return username;
     }
 
-    public void setAssistantType(SmartAssistantType assistantType) {
-        this.assistantType = assistantType;
+    public String getEmail() {
+        return email;
+    }
+
+    public Integer getQuestionNumber() {
+        return questionsNumber;
     }
 
     public Integer getRemainingQuestions() {
         return remainingQuestions;
     }
 
-    public void setRemainingQuestions(Integer remainingQuestions) {
-        this.remainingQuestions = remainingQuestions;
+    public AssistantType getAssistantType() {
+        return assistantType;
     }
 
     public Integer getRemainingQuestionsDelta() {
