+++ src/main/java/org/codedefenders/assistant/exceptions/GPTException.java
@@ -1,0 +1,11 @@
+package org.codedefenders.assistant.exceptions;
+
+/**
+ * This {@link Exception} represents an error occurred in the
+ * {@link org.codedefenders.assistant.services.GPTRequestDispatcher} while sending a request to OpenAI API.
+ */
+public class GPTException extends Exception{
+
+    public GPTException(){ }
+
+}
