+++ src/main/java/org/codedefenders/assistant/repositories/AssistantPromptRepository.java
@@ -1,4 +1,4 @@
-package org.codedefenders.persistence.database;
+package org.codedefenders.assistant.repositories;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -9,7 +9,7 @@ import javax.annotation.Nonnull;
 import javax.inject.Inject;
 
 import org.codedefenders.database.UncheckedSQLException;
-import org.codedefenders.model.AssistantPromptEntity;
+import org.codedefenders.assistant.entities.AssistantPromptEntity;
 import org.codedefenders.persistence.database.util.QueryRunner;
 import org.codedefenders.transaction.Transactional;
 import org.intellij.lang.annotations.Language;
@@ -18,6 +18,9 @@ import org.slf4j.LoggerFactory;
 
 import static org.codedefenders.persistence.database.util.ResultSetUtils.*;
 
+/**
+ * This repository provides methods for querying and updating the {@code assistant_prompts} table in the database.
+ */
 @Transactional
 public class AssistantPromptRepository {
 
@@ -29,6 +32,12 @@ public class AssistantPromptRepository {
         this.queryRunner = queryRunner;
     }
 
+    /**
+     * Maps a result set from the {@code assistant_prompts} table to an {@link AssistantPromptEntity} objet.
+     * @param rs the result set to map
+     * @return the {@link AssistantPromptEntity} corresponding to the result set
+     * @throws SQLException if a {@link SQLException} occurs while accessing the result set
+     */
     private AssistantPromptEntity assistantPromptEntityFromRS(ResultSet rs) throws SQLException {
         Integer id = rs.getInt("ID");
         String prompt = rs.getString("Prompt");
@@ -38,6 +47,10 @@ public class AssistantPromptRepository {
         return new AssistantPromptEntity(id, prompt, timestamp, asSeparateContext, defaultFlag);
     }
 
+    /**
+     * Gets the latest {@link AssistantPromptEntity}.
+     * @return the latest {@link AssistantPromptEntity}
+     */
     public Optional<AssistantPromptEntity> getLastPrompt() {
         @Language("SQL") String query = "SELECT ID, Prompt, Timestamp, As_separate_context, Default_flag " +
                 "FROM assistant_prompts ORDER BY Timestamp DESC LIMIT 1;";
@@ -49,6 +62,10 @@ public class AssistantPromptRepository {
         }
     }
 
+    /**
+     * Gets the latest {@link AssistantPromptEntity} marked as default.
+     * @return the latest {@link AssistantPromptEntity} marked as default
+     */
     public Optional<AssistantPromptEntity> getLastDefaultPrompt() {
         @Language("SQL") String query = "SELECT ID, Prompt, Timestamp, As_separate_context, Default_flag " +
                 "FROM assistant_prompts WHERE Default_flag = 1 ORDER BY Timestamp DESC LIMIT 1;";
@@ -60,6 +77,10 @@ public class AssistantPromptRepository {
         }
     }
 
+    /**
+     * Gets all {@link AssistantPromptEntity} ordered by timestamp, starting with the latest one.
+     * @return all {@link AssistantPromptEntity} ordered by timestamp, starting with the latest one
+     */
     public List<AssistantPromptEntity> getAllPrompts() {
         @Language("SQL") String query = "SELECT ID, Prompt, Timestamp, As_separate_context, Default_flag " +
                 "FROM assistant_prompts ORDER BY Timestamp DESC;";
@@ -71,12 +92,16 @@ public class AssistantPromptRepository {
         }
     }
 
-    public Optional<Integer> storePrompt(@Nonnull AssistantPromptEntity assistantPromptEntity) {
+    /**
+     * Stores a new {@link AssistantPromptEntity}.
+     * @param assistantPromptEntity the new {@link AssistantPromptEntity} to be stored
+     */
+    public void storePrompt(@Nonnull AssistantPromptEntity assistantPromptEntity) {
         @Language("SQL") String query = "INSERT INTO assistant_prompts "
                 + "(Prompt, As_separate_context, Default_flag) "
                 + "VALUES (?, ?, ?);";
         try {
-            return queryRunner.insert(query, resultSet -> nextFromRS(resultSet, rs -> rs.getInt(1)),
+            queryRunner.insert(query, resultSet -> nextFromRS(resultSet, rs -> rs.getInt(1)),
                     assistantPromptEntity.getPrompt(),
                     assistantPromptEntity.getAsSeparateContext(),
                     assistantPromptEntity.getDefaultFlag()
