+++ src/main/java/org/codedefenders/assistant/repositories/AssistantQuestionRepository.java
@@ -1,4 +1,4 @@
-package org.codedefenders.persistence.database;
+package org.codedefenders.assistant.repositories;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -9,7 +9,7 @@ import javax.annotation.Nonnull;
 import javax.inject.Inject;
 
 import org.codedefenders.database.UncheckedSQLException;
-import org.codedefenders.model.AssistantQuestionEntity;
+import org.codedefenders.assistant.entities.AssistantQuestionEntity;
 import org.codedefenders.persistence.database.util.QueryRunner;
 import org.codedefenders.transaction.Transactional;
 import org.intellij.lang.annotations.Language;
@@ -18,6 +18,9 @@ import org.slf4j.LoggerFactory;
 
 import static org.codedefenders.persistence.database.util.ResultSetUtils.*;
 
+/**
+ * This repository provides methods for querying and updating the {@code assistant_questions} table in the database.
+ */
 @Transactional
 public class AssistantQuestionRepository {
 
@@ -29,6 +32,12 @@ public class AssistantQuestionRepository {
         this.queryRunner = queryRunner;
     }
 
+    /**
+     * Maps a result set from the {@code assistant_questions} table to an {@link AssistantQuestionEntity} objet.
+     * @param rs the result set to map
+     * @return the {@link AssistantQuestionEntity} corresponding to the result set
+     * @throws SQLException if a {@link SQLException} occurs while accessing the result set
+     */
     private AssistantQuestionEntity assistantQuestionEntityFromRS(ResultSet rs) throws SQLException {
         Integer id = rs.getInt("ID");
         String question = rs.getString("Question");
@@ -42,6 +51,12 @@ public class AssistantQuestionRepository {
         return new AssistantQuestionEntity(id, question, answer, playerId, promptId, useful);
     }
 
+    /**
+     * Maps a result set containing dates and integers to a Java {@code map}.
+     * @param rs the result set to map
+     * @return the Java {@code map} corresponding to the result set
+     * @throws SQLException if a {@link SQLException} occurs while accessing the result set
+     */
     private Map<LocalDate, Integer> questionsAmountMapFromRS(ResultSet rs) throws SQLException {
         Map<LocalDate, Integer> map = new HashMap<>();
         while(rs.next()) {
@@ -52,7 +67,12 @@ public class AssistantQuestionRepository {
         return map;
     }
 
-    // add new question
+    /**
+     * Stores a new {@link AssistantQuestionEntity} containing the question text, the question prompt and the id of the
+     * player asking the question.
+     * @param assistantQuestionEntity the new {@link AssistantQuestionEntity} to be stored
+     * @return the id of the newly stored question
+     */
     public Optional<Integer> storeQuestion(@Nonnull AssistantQuestionEntity assistantQuestionEntity) {
         @Language("SQL") String query = "INSERT INTO assistant_questions "
                 + "(Question, Player_ID, Prompt_ID) "
@@ -69,7 +89,10 @@ public class AssistantQuestionRepository {
         }
     }
 
-    // update question with answer
+    /**
+     * Updates a question by storing an answer for it.
+     * @param assistantQuestionEntity the {@link AssistantQuestionEntity} containing the answer to be stored
+     */
     public void updateAnswer(@Nonnull AssistantQuestionEntity assistantQuestionEntity) {
         @Language("SQL") String query = "UPDATE assistant_questions " +
                 "SET answer = ? WHERE ID = ?";
@@ -87,7 +110,11 @@ public class AssistantQuestionRepository {
         }
     }
 
-    // get all questions with answer of given player
+    /**
+     * Gets all {@link AssistantQuestionEntity} of the given player containing valid answers, ordered by timestamp.
+     * @param playerId the id of the player who made the questions to retrieve
+     * @return a list of the questions with valid answers made by the player and ordered by timestamp
+     */
     public List<AssistantQuestionEntity> getQuestionsAndAnswersByPlayer(int playerId) {
         @Language("SQL") String query = "SELECT * FROM assistant_questions " +
                 "WHERE player_ID = ? AND answer IS NOT NULL ORDER BY timestamp;";
@@ -100,19 +127,28 @@ public class AssistantQuestionRepository {
         }
     }
 
-    // get amount of questions in the last X days with at least one question
-    public Map<LocalDate, Integer> getLastXAmountOfQuestions(int X) {
+    /**
+     * Starting from the current day and going backwards, this method retrieves the number of questions made in each day
+     * until it finds an amount of days with at least one questions equal to {@code numberOfDays}. The method returns
+     * all the days with at least one questions found in this way.
+     * @param numberOfDays the number of days with at least one question to be returned
+     * @return a map containing each day with at least one question and the amount of questions made in that day
+     */
+    public Map<LocalDate, Integer> getAmountOfQuestionsInTheLastNonEmptyDays(int numberOfDays) {
         @Language("SQL") String query = "SELECT DATE(Timestamp), count(*) FROM assistant_questions " +
                 "GROUP BY DATE(Timestamp) ORDER BY DATE(Timestamp) DESC LIMIT ?;";
         try {
-            return queryRunner.query(query, this::questionsAmountMapFromRS, X);
+            return queryRunner.query(query, this::questionsAmountMapFromRS, numberOfDays);
         } catch (SQLException e) {
             logger.error("SQLException while executing query", e);
             throw new UncheckedSQLException("SQLException while executing query", e);
         }
     }
 
-    // get total amount of questions (all players)
+    /**
+     * Gets the total amount of questions sent to the assistant.
+     * @return the total amount of questions sent to the assistant
+     */
     public Optional<Integer> getTotalQuestionsAmount() {
         @Language("SQL") String query = "SELECT COUNT(*) FROM assistant_questions;";
         try {
@@ -123,8 +159,14 @@ public class AssistantQuestionRepository {
         }
     }
 
-    // update usefulness of last the last question of given player
-    public void updateUsefulnessOfLastQuestionByPlayer(int playerId, boolean usefulness) {
+    /**
+     * Updates the latest question sent by the given player storing a feedback for it.
+     * @param playerId the id of the player who made the question
+     * @param usefulness {@code true} if the question has been useful; {@code false} otherwise
+     * @return {@code true} if the update was applied successfully; {@code false} if no question to update was found for
+     * the given player
+     */
+    public boolean updateUsefulnessOfLastQuestionByPlayer(int playerId, boolean usefulness) {
         @Language("SQL") String query = "UPDATE assistant_questions, " +
                 "(SELECT id FROM assistant_questions WHERE player_id = ? ORDER BY timestamp DESC LIMIT 1) " +
                 "AS last_player_question " +
@@ -133,12 +175,13 @@ public class AssistantQuestionRepository {
         try {
             int updatedRows = queryRunner.update(query, playerId, usefulness);
             if (updatedRows != 1) {
-                throw new UncheckedSQLException("Couldn't update assistant answer usefulness");
+                return false;
             }
         } catch (SQLException e) {
             logger.error("SQLException while executing query", e);
             throw new UncheckedSQLException("SQLException while executing query", e);
         }
+        return true;
     }
 
 }
