+++ src/main/java/org/codedefenders/assistant/repositories/AssistantUserSettingsRepository.java
@@ -1,4 +1,4 @@
-package org.codedefenders.persistence.database;
+package org.codedefenders.assistant.repositories;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
@@ -8,8 +8,8 @@ import java.util.Optional;
 import javax.inject.Inject;
 
 import org.codedefenders.database.UncheckedSQLException;
-import org.codedefenders.model.AssistantUserSettingsEntity;
-import org.codedefenders.model.SmartAssistantType;
+import org.codedefenders.assistant.entities.AssistantUserSettingsEntity;
+import org.codedefenders.assistant.entities.AssistantType;
 import org.codedefenders.persistence.database.util.QueryRunner;
 import org.codedefenders.transaction.Transactional;
 import org.intellij.lang.annotations.Language;
@@ -18,6 +18,9 @@ import org.slf4j.LoggerFactory;
 
 import static org.codedefenders.persistence.database.util.ResultSetUtils.*;
 
+/**
+ * This repository provides methods for querying and updating the {@code assistant_user_settings} table in the database.
+ */
 @Transactional
 public class AssistantUserSettingsRepository {
 
@@ -29,6 +32,12 @@ public class AssistantUserSettingsRepository {
         this.queryRunner = queryRunner;
     }
 
+    /**
+     * Maps a result set from the {@code assistant_user_settings} table to an {@link AssistantUserSettingsEntity} objet.
+     * @param rs rs the result set to map
+     * @return the {@link AssistantUserSettingsEntity} corresponding to the result set
+     * @throws SQLException if a {@link SQLException} occurs while accessing the result set
+     */
     private AssistantUserSettingsEntity assistantUserSettingsEntityFromRS(ResultSet rs) throws SQLException {
         Integer id = rs.getInt("ID");
         if(rs.wasNull()) {
@@ -44,10 +53,13 @@ public class AssistantUserSettingsRepository {
             assistantType = "NONE";
         }
         return new AssistantUserSettingsEntity(id, userId, username, email, questionsNumber, remainingQuestions,
-                SmartAssistantType.valueOf(assistantType));
+                AssistantType.valueOf(assistantType));
     }
 
-    // get all users settings
+    /**
+     * Gets the {@link AssistantUserSettingsEntity} for each active user.
+     * @return a list containing the {@link AssistantUserSettingsEntity} for each active user
+     */
     public List<AssistantUserSettingsEntity> getAllAssistantUserSettings() {
         @Language("SQL") String query = "SELECT a.ID, u.User_ID, u.Username, u.Email, " +
                 "a.Questions_number, a.Remaining_questions, a.Assistant_type " +
@@ -63,7 +75,11 @@ public class AssistantUserSettingsRepository {
         }
     }
 
-    // get settings of given user
+    /**
+     * Gets the {@link AssistantUserSettingsEntity} of a given user.
+     * @param userId the id of the given user
+     * @return the {@link AssistantUserSettingsEntity} of the given user
+     */
     public Optional<AssistantUserSettingsEntity> getAssistantUserSettingsByUserId(int userId) {
         @Language("SQL") String query = "SELECT a.ID, u.User_ID, u.Username, u.Email, " +
                 "a.Questions_number, a.Remaining_questions, a.Assistant_type " +
@@ -78,11 +94,16 @@ public class AssistantUserSettingsRepository {
         }
     }
 
-    // insert or update settings of given user
+    /**
+     * Updates the {@link AssistantUserSettingsEntity} of a specific user. If no settings are present in the database
+     * for that user, then a new row is inserted for the user and is updated with the given settings.
+     * @param settings the {@link AssistantUserSettingsEntity} to be updated
+     * @return the id of the updated {@link AssistantUserSettingsEntity}
+     */
     public Optional<Integer> updateAssistantUserSettings(AssistantUserSettingsEntity settings) {
         int userId = settings.getUserId();
         int remainingQuestionsDelta = settings.getRemainingQuestionsDelta();
-        SmartAssistantType assistantType = settings.getAssistantType();
+        AssistantType assistantType = settings.getAssistantType();
         @Language("SQL") String query = "INSERT INTO assistant_user_settings (User_ID, Remaining_questions, Assistant_type) " +
                 "VALUES (?, ?, ?) " +
                 "ON DUPLICATE KEY UPDATE Remaining_questions = Remaining_questions + ?, Assistant_type = ?;";
