--- src/main/java/org/codedefenders/model/SmartAssistantType.java
@@ -1,5 +1,0 @@
-package org.codedefenders.model;
-
-public enum SmartAssistantType {
-    NONE, NOT_GUIDED
-}
