+++ src/main/java/org/codedefenders/servlets/games/battleground/MultiplayerGameManager.java
@@ -82,7 +82,7 @@ import org.codedefenders.servlets.games.GameManagingUtils;
 import org.codedefenders.servlets.games.GameProducer;
 import org.codedefenders.servlets.util.Redirect;
 import org.codedefenders.servlets.util.ServletUtils;
-import org.codedefenders.smartassistant.AssistantService;
+import org.codedefenders.assistant.services.AssistantService;
 import org.codedefenders.util.Constants;
 import org.codedefenders.util.Paths;
 import org.codedefenders.util.URLUtils;
