--- src/main/java/org/codedefenders/smartassistant/GPTObjects/GPTCompletionUsage.java
@@ -1,7 +1,0 @@
-package org.codedefenders.smartassistant.GPTObjects;
-
-public class GPTCompletionUsage {
-    private Integer promptTokens;
-    private Integer completionTokens;
-    private Integer totalTokens;
-}
