--- src/main/java/org/codedefenders/smartassistant/GPTObjects/GPTRole.java
@@ -1,9 +1,0 @@
-package org.codedefenders.smartassistant.GPTObjects;
-
-import com.google.gson.annotations.SerializedName;
-
-public enum GPTRole {
-    @SerializedName("system") SYSTEM,
-    @SerializedName("user") USER,
-    @SerializedName("assistant") ASSISTANT
-}
