--- src/main/java/org/codedefenders/smartassistant/exceptions/GPTException.java
@@ -1,11 +1,0 @@
-package org.codedefenders.smartassistant.exceptions;
-
-public class GPTException extends Exception{
-
-    public GPTException(){ }
-
-    public GPTException(String message) {
-        super(message);
-    }
-
-}
